Rename mentionbase => base
