+++ packages/nimble-components/src/rich-text-mention/base/models/mapping-config.ts
