+++ packages/nimble-components/src/rich-text-mention/base/models/mention-internals.ts
