+++ packages/nimble-components/src/rich-text-mention/base/models/rich-text-mention-validator.ts
